requirements.txt
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
yfinance==0.2.44
pyportfolioopt==1.5.5
plotly==5.24.1
streamlit==1.39.0
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.9.2
fastapi==0.115.0
uvicorn==0.30.6
duckdb==1.1.3
tqdm==4.66.5
fredapi==0.5.0
feedparser==6.0.11
sec-edgar-downloader==5.0.2
python-dateutil==2.9.0.post0
